Move integration tests to a test folder
As correctly suggested by @ZER0 we should move the integration
tests to a `tests` folder next to `src` so cargo does compile and
run them.

That allows us to see how the API looks like and therefore prevent
us to push anythig that does not look good related to the API shape.
