+++ src/range.rs
@@ -191,39 +191,6 @@ mod tests {
     use super::*;
     use anyhow::{Error, Result};
 
-    fn max_bound_gadget(
-        composer: &mut StandardComposer,
-        max_range: BlsScalar,
-        witness: BlsScalar,
-        result: bool,
-    ) {
-        let witness = AllocatedScalar::allocate(composer, witness);
-        let (res, _) = max_bound(composer, max_range, witness);
-
-        let mut outcome = BlsScalar::zero();
-        if result {
-            outcome = BlsScalar::one()
-        }
-        composer.constrain_to_constant(res, outcome, BlsScalar::zero());
-    }
-
-    fn range_check_gadget(
-        composer: &mut StandardComposer,
-        max_range: BlsScalar,
-        min_range: BlsScalar,
-        witness: BlsScalar,
-        result: bool,
-    ) {
-        let witness = AllocatedScalar::allocate(composer, witness);
-        let res = range_check(composer, min_range, max_range, witness);
-
-        let mut outcome = BlsScalar::zero();
-        if result {
-            outcome = BlsScalar::one()
-        }
-        composer.constrain_to_constant(res, outcome, BlsScalar::zero());
-    }
-
     #[test]
     fn counting_scalar_bits() {
         assert_eq!(bits_count(BlsScalar::zero()), 1);
@@ -231,220 +198,6 @@ mod tests {
         assert_eq!(bits_count(BlsScalar::from(3u64)), 2);
         let two_pow_128 = BlsScalar::from(2u64).pow(&[128u64, 0, 0, 0]);
         assert_eq!(bits_count(two_pow_128), 129);
-    }
-
-    #[test]
-    fn test_maybe_equal() -> Result<(), Error> {
-        // Generate Composer & Public Parameters
-        let pub_params = PublicParameters::setup(1 << 10, &mut rand::thread_rng())?;
-        let (ck, vk) = pub_params.trim(1 << 9)?;
-
-        let is_equal_gadget =
-            |composer: &mut StandardComposer, num_1: u64, num_2: u64, result: bool| {
-                let a = AllocatedScalar::allocate(composer, BlsScalar::from(num_1));
-                let b = AllocatedScalar::allocate(composer, BlsScalar::from(num_2));
-
-                let bit = maybe_equal(composer, a, b);
-
-                let mut outcome = BlsScalar::zero();
-                if result {
-                    outcome = BlsScalar::one()
-                }
-                composer.constrain_to_constant(bit, outcome, BlsScalar::zero());
-            };
-
-        // Proving
-        // Should pass as 100 == 100
-        let mut prover = Prover::new(b"testing");
-        is_equal_gadget(prover.mut_cs(), 100, 100, true);
-
-        prover.preprocess(&ck)?;
-        let proof = prover.prove(&ck)?;
-
-        // Verification
-        let mut verifier = Verifier::new(b"testing");
-        is_equal_gadget(verifier.mut_cs(), 0, 0, true);
-
-        verifier.preprocess(&ck).expect("Preprocessing error");
-        assert!(verifier
-            .verify(&proof, &vk, &vec![BlsScalar::zero()])
-            .is_ok());
-
-        // Proving
-        // Should fail as 20 != 3330
-        let mut prover = Prover::new(b"testing");
-        is_equal_gadget(prover.mut_cs(), 20, 3330, false);
-
-        prover.preprocess(&ck)?;
-        let proof = prover.prove(&ck)?;
-
-        // Verification
-        let mut verifier = Verifier::new(b"testing");
-        is_equal_gadget(verifier.mut_cs(), 0, 0, false);
-
-        verifier.preprocess(&ck)?;
-        assert!(verifier
-            .verify(&proof, &vk, &vec![BlsScalar::zero()])
-            .is_ok());
-
-        Ok(())
-    }
-
-    #[test]
-    fn max_bound_test() -> Result<(), Error> {
-        // Generate Composer & Public Parameters
-        let pub_params = PublicParameters::setup(1 << 11, &mut rand::thread_rng())?;
-        let (ck, vk) = pub_params.trim(1 << 10)?;
-
-        struct TestCase {
-            max_range: BlsScalar,
-            witness: BlsScalar,
-            expected_result: bool,
-        }
-        let test_cases = vec![
-            TestCase {
-                max_range: BlsScalar::from(2u64).pow(&[128u64, 0, 0, 0]) - BlsScalar::one(),
-                witness: BlsScalar::from(2u64).pow(&[127u64, 0, 0, 0]),
-                expected_result: true,
-            },
-            TestCase {
-                max_range: BlsScalar::from(200u64),
-                witness: BlsScalar::from(100u64),
-                expected_result: true,
-            },
-            TestCase {
-                max_range: BlsScalar::from(100u64),
-                witness: BlsScalar::from(200u64),
-                expected_result: false,
-            },
-            TestCase {
-                max_range: BlsScalar::from(2u64).pow(&[128u64, 0, 0, 0]) - BlsScalar::one(),
-                witness: BlsScalar::from(2u64).pow(&[130u64, 0, 0, 0]),
-                expected_result: false,
-            },
-        ];
-
-        for case in test_cases.into_iter() {
-            // Proving
-            let mut prover = Prover::default();
-
-            max_bound_gadget(
-                prover.mut_cs(),
-                case.max_range,
-                case.witness,
-                case.expected_result,
-            );
-            prover.preprocess(&ck)?;
-            let proof = prover.prove(&ck)?;
-
-            // Verification
-            let mut verifier = Verifier::default();
-            max_bound_gadget(
-                verifier.mut_cs(),
-                case.max_range,
-                case.witness,
-                case.expected_result,
-            );
-            verifier.preprocess(&ck)?;
-            assert!(verifier
-                .verify(&proof, &vk, &vec![BlsScalar::zero()])
-                .is_ok());
-        }
-        Ok(())
-    }
-    #[test]
-    fn range_check_test() -> Result<(), Error> {
-        // Generate Composer & Public Parameters
-        let pub_params = PublicParameters::setup(1 << 11, &mut rand::thread_rng())?;
-        let (ck, vk) = pub_params.trim(1 << 10)?;
-
-        struct TestCase {
-            max_range: BlsScalar,
-            min_range: BlsScalar,
-            witness: BlsScalar,
-            expected_result: bool,
-        }
-        let test_cases = vec![
-            TestCase {
-                min_range: BlsScalar::from(50_000u64),
-                max_range: BlsScalar::from(250_000u64),
-                witness: BlsScalar::from(50_001u64),
-                expected_result: true,
-            },
-            TestCase {
-                min_range: BlsScalar::from(50_000u64),
-                max_range: BlsScalar::from(250_000u64),
-                witness: BlsScalar::from(250_001u64),
-                expected_result: false,
-            },
-            TestCase {
-                min_range: BlsScalar::from(50_000u64),
-                max_range: BlsScalar::from(250_000u64),
-                witness: BlsScalar::from(250_000u64),
-                expected_result: false,
-            },
-            TestCase {
-                min_range: BlsScalar::from(50_000u64),
-                max_range: BlsScalar::from(250_000u64),
-                witness: BlsScalar::from(249_000u64),
-                expected_result: true,
-            },
-            TestCase {
-                min_range: BlsScalar::from(50_000u64),
-                max_range: BlsScalar::from(250_000u64),
-                witness: BlsScalar::from(50_000u64),
-                expected_result: true,
-            },
-            TestCase {
-                min_range: BlsScalar::from(50_000u64),
-                max_range: BlsScalar::from(250_000u64),
-                witness: BlsScalar::from(49_999u64),
-                expected_result: false,
-            },
-            TestCase {
-                min_range: BlsScalar::from(2u64).pow(&[126u64, 0, 0, 0]),
-                max_range: BlsScalar::from(2u64).pow(&[127u64, 0, 0, 0]) + BlsScalar::one(),
-                witness: BlsScalar::from(2u64).pow(&[127u64, 0, 0, 0]) - BlsScalar::one(),
-                expected_result: true,
-            },
-            TestCase {
-                min_range: BlsScalar::from(50_000u64),
-                max_range: BlsScalar::from(250_000u64),
-                witness: BlsScalar::from(18_598u64),
-                expected_result: false,
-            },
-        ];
-
-        for case in test_cases.into_iter() {
-            // Proving
-            let mut prover = Prover::default();
-
-            range_check_gadget(
-                prover.mut_cs(),
-                case.max_range,
-                case.min_range,
-                case.witness,
-                case.expected_result,
-            );
-            prover.preprocess(&ck)?;
-            let proof = prover.prove(&ck)?;
-
-            // Verification
-            let mut verifier = Verifier::default();
-            range_check_gadget(
-                verifier.mut_cs(),
-                case.max_range,
-                case.min_range,
-                case.witness,
-                case.expected_result,
-            );
-            verifier.preprocess(&ck)?;
-            assert!(verifier
-                .verify(&proof, &vk, &vec![BlsScalar::zero()])
-                .is_ok());
-        }
-
-        Ok(())
     }
 
     #[test]
